# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/announcement/create": {
            "post": {
                "tags": [
                    "Announcement"
                ],
                "summary": "Create a new announcement",
                "description": "Create a new announcement.",
                "operationId": "36b5cd0f70fa61d1b76caf65e4905a86",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "New Announcement"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "This is a new announcement"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-04-01"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "New York"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Public"
                                    },
                                    "required_skills": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Skill 1",
                                            "Skill 2"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Announcement created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Announcement Created Successfully !"
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "session": {
                                            "type": "string",
                                            "example": "admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/announcement/update/{announcement}": {
            "put": {
                "tags": [
                    "Announcement"
                ],
                "summary": "Update an existing announcement",
                "description": "Update an existing announcement.",
                "operationId": "a074ad6e13a3359f28343ee861621f45",
                "parameters": [
                    {
                        "name": "announcement",
                        "in": "path",
                        "description": "ID of the announcement to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Announcement"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "This is the updated announcement"
                                    },
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-04-01"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Updated Location"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Updated Type"
                                    },
                                    "required_skills": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Updated Skill 1",
                                            "Updated Skill 2"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Announcement updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Announcement data updated successfully"
                                        },
                                        "updated_announcement": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Announcement not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/announcement/delete/{announcement}": {
            "delete": {
                "tags": [
                    "Announcement"
                ],
                "summary": "Delete an existing announcement",
                "description": "Delete an existing announcement.",
                "operationId": "94f5438c8195b1e7535ad7398f01962b",
                "parameters": [
                    {
                        "name": "announcement",
                        "in": "path",
                        "description": "ID of the announcement to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Announcement deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Announcement 'Title' deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Announcement not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/announcements/all": {
            "get": {
                "tags": [
                    "Announcement"
                ],
                "summary": "Retrieve all announcements",
                "description": "Retrieve all announcements.",
                "operationId": "7c3790258e759d8c432b8381841f0d5e",
                "responses": {
                    "200": {
                        "description": "List of all announcements",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "announcements": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/api/announcements/filter": {
            "post": {
                "tags": [
                    "Announcement"
                ],
                "summary": "Filter announcements by keyword",
                "description": "Filter announcements by keyword.",
                "operationId": "5fe09ff1a0c9c296ab5a543418405330",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "keyword": {
                                        "type": "string",
                                        "example": "Event"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "List of filtered announcements",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "announcements": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
    }
}